requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
pydantic>=2.6
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

PyPDF2>=3.0
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1

email-validator
python-multipart>=0.0.9
